Bus timeout no longer supports block-forever
Calling recv() on the Redis bus with an indefinite timeout leads to
scenarios where signals cannot be handled, and should generally be
avoided.  Related, the Timer can no longer return a negative amount for
time remaining, which was leading to unexpected block-forever scenarios
on recv().  Instead, if a client needs to wait essentially forever, then
wrap the recv() in a loop.

+++ evergreen/src/common/auth.rs
@@ -9,7 +9,7 @@ use eg::{Client, Editor, EgError, EgEvent, EgResult, EgValue};
 use md5;
 use std::fmt;
 
-const LOGIN_TIMEOUT: i32 = 30;
+const LOGIN_TIMEOUT: u64 = 30;
 
 // Default time for extending a persistent session: ten minutes
 const DEFAULT_RESET_INTERVAL: i32 = 10 * 60;
